third_party/blink/web_tests/virtual/stylable-select-disabled/fast/forms/select/stylable-select/nested-select-console-message-expected.txt

CONSOLE ERROR: A <select> tag was parsed within another <select> tag and was converted into </select>. This behavior will change in a future browser version. Please add the missing </select> end tag.
option
